test_project/dev_requirements.txt
coverage==7.6.4
python-dotenv==1.0.1
